requirements.txt
asgiref==3.7.2
Django==3.2
django-colorfield==0.9.0
djangorestframework==3.12.4
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
